requirements.txt
pytest==8.3.5
selenium==4.38.0
allure-pytest==2.15.0
requests==2.32.3
